# Deprecate obsolete pattern matching functions in std.string