These functions are old D1 string functions from before std.regex's existence. With std.regex and std.algorithm, these functions are no longer needed, and their use should be discouraged, as the std.algorithm versions will be faster (better search algorithms, better handling of auto-decoding).

Plus, this removes a domain specific language that's in use only in one minor part of D.