@JackStouffer 2017-05-08
jeez `std.xml` is messy

Seeing as we already have a xml replacement in the pipe I'm tempted to create a local copy of `munch` and call it a day.