@JackStouffer 2017-05-08
Ok, took a bit longer than expected, but I fixed `std.xml`