@DmitryOlshansky 2017-05-08
Starting and stopping time around such a trivial operation is going to
easily hide the real latency of operation due to coarse granularity of the
timer.
So please try to measure the full loop and make sure the compiler doesn't
optimize it away.

2017-05-08 16:54 GMT+02:00 Jack Stouffer <notifications@github.com>:

> Ha, holy hell. I was worried that the countUntil method might actually be
> slower than munch. Turns out it's 300x faster.
>
> # old
> $ ldc2 -O -release test.d && ./test
> Total time: 10 secs, 695 ms, 464 μs, and 8 hnsecs
>
> # new
> $ ldc2 -O -release test.d && ./test
> Total time: 30 ms, 100 μs, and 6 hnsecs
>
> code
>
> import std.stdio;import std.conv;import std.range;import std.string;import std.ascii;import std.algorithm.searching;import std.datetime;import std.utf;
> auto func(T)(ref T s)
> {
>     auto i = s.byCodeUnit.countUntil!(a => !isWhite(a));
>     auto ret = s[0 .. i];
>     s = s[i .. $];
>     return ret;
> }
> __gshared char[] test = [' ',' ',' ',' ',' ',' ','h','t','m','l'];
> void main()
> {
>     enum n = 100_000_000;
>
>     char[] result;
>     StopWatch sw;
>     Duration sum;
>     TickDuration last = TickDuration.from!"seconds"(0);
>
>     foreach(i; 0 .. n)
>     {
>         sw.start();
>
>         // New
>         result = func(test);
>
>         // Old
>         //result = munch(test, whitespace);
>
>         sw.stop();
>
>         test = [' ',' ',' ',' ',' ',' ','h','t','m','l'];
>         auto time = sw.peek() - last;
>         last = sw.peek();
>
>         sum += time.to!Duration();
>     }
>
>     writeln("Total time: ", sum);
> }
>
> —
> You are receiving this because you are subscribed to this thread.
> Reply to this email directly, view it on GitHub
> <https://github.com/dlang/phobos/pull/5378#issuecomment-299889999>, or mute
> the thread
> <https://github.com/notifications/unsubscribe-auth/AARMqrlfld38rvn_YkX4YeYHZ1DoXiw0ks5r3yypgaJpZM4NTWKC>
> .
>