@JackStouffer 2017-05-08
Ok, I don't want to turn this into a discussion about proper benchmarking code so I deleted that comment.

Let's instead focus on the merit of the deprecations.