@wilzbach 2017-05-08
> Let's instead focus on the merit of the deprecations.

Ok, this definitely requires the approval of @andralex.
I know too little about the history of std.string to be able to judge this, but they definitely look old as e.g. `removechars` would be named `removeChars` or `removePattern` nowadays.
However especially from the PoV that we want to get rid of auto-decoding this seams to me a step in the right direction!