@andralex 2017-05-09
On the plus side, we improve consistency of the stdlib API and we reduce its code size (albeit by a small amount). On the minus side, we deprecate existing uses without a simple migration path. Consider someone who's not the author of the project but a user. They'd need to start doing nontrivial code surgery on a foreign project to use regexes instead of patterns etc.

For this kind of evolution I want us to define a path for marginalizing/obsoleting the code without removing it. Mark it in a specific way in the documentation, later make it only accessible if you click on "See obsolete APIs", etc. That way new code will not use it but old code will continue to work. Some libraries do this with good results. cc @MartinNowak @CyberShadow @WalterBright