@JackStouffer 2017-05-09
>Mark it in a specific way in the documentation, later make it only accessible if you click on "See obsolete APIs", etc.

Isn't this why undeaD exists?