@CyberShadow 2017-05-09
+1 for moving these to undeaD.

FWIW an anecdotal data point: I do a lot of text processing in D and I've never used these functions.