@jmdavis 2017-05-09
As to the history, a lot of the std.string functions were in D1 and had names which did not match Phobos' current naming schemes, and several years ago, when I fixed the names in std.string to follow Phobos' current naming scheme, I did not fix the pattern functions, because the idea was that we were going to replace them with functions with the proper capitalization that used regexes instead of the patterns (and if we were going to do that, it made no sense to make folks change their code to use the new capitalization and _then_ make them change their code to use std.regex). However, no one has ever done the work to create the new versions that use std.regex, so they've just sat there.

I'm all for moving them to undead, but if we do that, we'll need to go through the normal deprecation process for them in Phobos, not simply remove them. And if we remove them, it would be nice to replace them with versions that use std.regex, but someone would have to do that work, and I don't know how much use these functions get (I've never used them).