@jmdavis 2017-05-09
> I think you'd be hard pressed to find something these functions can do that can't already be done with std.regex + std.algorithm

Then there isn't much point in replacing them in std.string.