@JackStouffer 2017-05-09
>Then there isn't much point in replacing them in std.string.

Exactly?

Maybe we're talking past each other, but I'm not sure if you're for or against having these functions go through the deprecated -> undocumented -> removed -> undeaD cycle.