@andralex 2017-05-09
The way I see evolution is marginalize -> undocument -> deprecate -> move to undead.